Moved variable class
